Add pre-made sudoku puzzles, and their solutions.
The puzzles and their solutions are borrowed from a 2017 dataset of
1 million sudoku puzzles posted by Kyubyong Park on the kaggle.com site.
The web url for this dataset (at the time of writing) is:
https://www.kaggle.com/datasets/bryanpark/sudoku

The puzzles used in this csv file are only a portion of the dataset.

The csv file includes puzzles on the left, and solutions on the right.

The first nine numbers (in either the puzzle or the solution) correspond
to the first row of the Sudoku, the next nine to the second row, and
so on.

In this dataset, empty cells in the puzzle are represented by a `0`.