Fix hints bug. Closes issue #2
Logic has been re-structured to support hints without any peculiar
behaviour.

After the un-solved and solved grids are constructed, a new list that
keeps track of the locations of unfilled cells is also constructed.

A hint is now a matter of choosing a random location from this
new list, and revealing the correct number in this same location.

The list of unfilled cells is updated accordingly, after a hint has been
given.